% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerater.R
\name{dataGenerater}
\alias{dataGenerater}
\title{Generate a mutated FASTA File and List Variable based on Healthy Version}
\usage{
dataGenerater(data, nummut)
}
\description{
Generate a mutated FASTA File and List Variable based on Healthy Version
}
