% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImporter.R
\name{fastaReader}
\alias{fastaReader}
\title{Import and Check the FASTA DNA file}
\usage{
fastaReader(filename)
}
\arguments{
\item{filename}{The file name of the FASTA file wish to be read.}
}
\value{
A Base Characters contained in the sequence. (In real case,
the number of returned items would equal to the length of the input
DNA sequence if sequence is valid.)
}
\description{
A function that reads in FASTA files and check if valid.
}
\examples{
# Example 1:
# Import data in sample raw FASTA format
# Note: to read FASTA files, it requires installation of seqinr R package
library("seqinr")
inputhea <- system.file("extdata", "sample.fasta", package = "oncoAnalysis")
inputmut <- system.file("extdata", "samplemut.fasta", package = "oncoAnalysis")

# Read using fastaReader()
sampleseq <- fastaReader(inputhea)
samplemutseq <- fastaReader(inputmut)

}
\references{
Charif, D. and Lobry, J.R. (2007). SeqinR 1.0-2: a contributed package to the
R project for statistical computing devoted to biological sequences retrieval
and analysis

mikemike, et al.
“Test If Characters Are in a String.” Stack Overflow, 12 Apr. 2012,
\href{https://stackoverflow.com/questions/10128617/test-if-characters-are-in-a-string}{Link}.

user2588829, et al.
“Break/Exit Script.” Stack Overflow, 24 July 2013,
\href{https://stackoverflow.com/questions/17837289/break-exit-script}{Link}.
}
