% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoAnalysisPlot.R
\name{mutCompPlot}
\alias{mutCompPlot}
\title{Visualize the Base number comparison between Healthy and Mutated Sequence.}
\usage{
mutCompPlot(datahea, datamut)
}
\arguments{
\item{datahea}{A vector containing DNA sequence(healthy)}

\item{datamut}{A vector containing DNA sequence(suspect mutated)}
}
\value{
Returns a plot of mutation details.
}
\description{
A function that visualize the base number comparison between healthy
and mutated Sequence.
}
\examples{
# Using sampleseq and samplemutseq dataset available with package
# Read in data
load("./data/sampleseq.rda")
load("./data/samplemutseq.rda")
sampleseq1 <- sampleseq[1:(length(sampleseq))]
samplemutseq1 <- samplemutseq[1:(length(samplemutseq))]

# Compare the base numbers in Sequence.
mutCompPlot(sampleseq1, samplemutseq1)

}
\references{
Holtz, Yan. “Basic Histogram with GGPLOT2.” – The R Graph Gallery.
\href{https://r-graph-gallery.com/220-basic-ggplot2-histogram.html}{Link}

“Modify Components of a Theme - Theme.” - Theme • ggplot2
\href{https://ggplot2.tidyverse.org/reference/theme.html}{Link}

“Bar Charts - geom_bar.” - geom_bar • ggplot2
\href{https://ggplot2.tidyverse.org/reference/geom_bar.html}{Link}

Chang, Winston. “R Graphics Cookbook, 2nd Edition.”
3.9 Adding Labels to a Bar Graph, 11 Nov. 2022
\href{https://r-graphics.org/recipe-bar-graph-labels}{Link}

AndrewAndrew 1322 silver badges33 bronze badges, and jtr13jtr13
\enumerate{
\item “Ggplot geom_col Legend Not Showing.” Stack Overflow, 1 Mar. 1965.
\href{https://stackoverflow.com/questions/48134472/ggplot-geom-col-legend-not-showing}{Link}
}
}
