% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutChecker.R
\name{mutTable}
\alias{mutTable}
\title{Display the Mutation Matrix generated by mutChecker in table format}
\usage{
mutTable(mutatmatrix)
}
\arguments{
\item{mutatmatrix}{A mutation matrix generated by mutChecker}
}
\value{
Displays the mutation table.
\itemize{
\item A table containing mutation details. (Column names represent the
base in original sequence, Role names represent the base in
mutated sequence. Used to check for example how many base A has mutated
into base T.)
}
}
\description{
A function that displays Mutation Matrix in table format
}
\examples{
# Example 1:
# Using sampleseq and samplemutseq dataset available with package
# Read in data
sampleseq1 <- sampleseq[1:(length(sampleseq))]
samplemutseq1 <- samplemutseq[1:(length(samplemutseq))]

# Check the number of mutations.
mutCheckerResults <- mutChecker(
                          sampleseq1,
                          samplemutseq1)
mutTable(mutCheckerResults$MutMatrix)

}
\references{
Coghlan, Avril. "Little book of R for Bioinformatics."
Cambridge, UK: Welcome Trust Sanger Institute (2011).
\href{https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/index.html}{Link}.

“How to Create Tables in R?” GeeksforGeeks, 27 Dec. 2021.
\href{https://www.geeksforgeeks.org/how-to-create-tables-in-r/}{Link}.
}
