% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplefalseseq}
\alias{samplefalseseq}
\title{Data set containing Sample DNA Sequences (Incorrect formatted)}
\format{
A vector containing strings, each string contains one base
\describe{
\item{Base.1}{A DNA base}
\item{Base.2}{A DNA base}
\item{Base.N}{A DNA base}
}
}
\source{
National Library of Medicine, and self generated.
}
\usage{
samplefalseseq
}
\description{
A Incorrect formatted Sample DNA Sequences
}
\examples{
\dontrun{
 samplefalseseq
}
}
\keyword{datasets}
