% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplemutseq}
\alias{samplemutseq}
\title{Data set containing Sample DNA Sequences (Mutated)}
\format{
A vector containing strings, each string contains one base
\describe{
\item{Base.1}{A DNA base}
\item{Base.2}{A DNA base}
\item{Base.N}{A DNA base}
}
}
\source{
National Library of Medicine, and self generated.
}
\usage{
samplemutseq
}
\description{
A Mutated Sample DNA Sequences(P53)
}
\examples{
\dontrun{
 samplemutseq
}
}
\keyword{datasets}
