% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleseq}
\alias{sampleseq}
\title{Data set containing Sample DNA Sequences (Healthy)}
\format{
A vector containing strings, each string contains one base
\describe{
\item{Base.1}{A DNA base}
\item{Base.2}{A DNA base}
\item{Base.N}{A DNA base}
}
}
\source{
National Library of Medicine.
}
\usage{
sampleseq
}
\description{
A Healthy Sample DNA Sequences(P53)
}
\examples{
\dontrun{
 sampleseq
}
}
\keyword{datasets}
