% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutChecker.R
\name{mutChecker}
\alias{mutChecker}
\title{Check if Mutation Exists along with its numbers}
\usage{
mutChecker(datahea, datamut)
}
\arguments{
\item{datahea}{A vector containing DNA sequence(healthy)}

\item{datamut}{A vector containing DNA sequence(suspect mutated)}
}
\value{
Returns a list containing mutation details.
\itemize{
\item The total number of mutations.
\item The number of Base A mutated.
\item The number of Base T mutated.
\item The number of Base C mutated.
\item The number of Base G mutated.
\item The position of each mutation.
\item The matrix containing mutation details. (Column names represent the
base in original sequence, Role names represent the base in
mutated sequence. Used to check for example how many base A has mutated
into base T.)
}
}
\description{
A function that checks if there is mutation occurred
}
\examples{
# Example 1:
# Using sampleseq and samplemutseq dataset available with package
\dontrun{
# Read in data
load("./data/sampleseq.rda")
load("./data/samplemutseq.rda")
sampleseq1 <- sampleseq[1:(length(sampleseq))]
samplemutseq1 <- samplemutseq[1:(length(samplemutseq))]

# Check the number of mutations.
mutCheckerResults <- mutChecker(
                          sampleseq1,
                          samplemutseq1)
}

# Example 2:
# Import data in sample raw FASTA format
# Note: to read FASTA files, it requires installation of seqinr R package
library("seqinr")
inputhea <- system.file("extdata", "sample.fasta", package = "oncoAnalysis")
inputmut <- system.file("extdata", "samplemut.fasta", package = "oncoAnalysis")

# Read using read.fasta()
sampleseq <- seqinr::read.fasta(file = inputhea)
samplemutseq <- seqinr::read.fasta(file = inputmut)

# Process and store data
sampleseq1 <- sampleseq$Sample[1:(length(sampleseq$Sample))]
samplemutseq1 <- samplemutseq$Samplemut[1:(length(samplemutseq$Samplemut))]

# Check the number of mutations.
mutCheckerResults <- mutChecker(
                          sampleseq1,
                          samplemutseq1)

}
\references{
Charif, D. and Lobry, J.R. (2007). SeqinR 1.0-2: a contributed package to the
R project for statistical computing devoted to biological sequences retrieval
and analysis
\href{https://cran.r-project.org/web/packages/seqinr/index.html}{Link}

Coghlan, Avril. "Little book of R for Bioinformatics."
Cambridge, UK: Welcome Trust Sanger Institute (2011).
\href{https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/index.html}{Link}.

“R Compare Vectors & Find Differences (5 Examples):
Identify Similarities.” Statistics Globe, 18 Mar. 2022.
\href{https://statisticsglobe.com/compare-vectors-and-find-differences-in-r}{Link}.

“R For Loop (with Examples).” DataMentor, 8 Oct. 2018.
\href{https://www.datamentor.io/r-programming/for-loop/}{Link}.

“How to Create Tables in R?” GeeksforGeeks, 27 Dec. 2021.
\href{https://www.geeksforgeeks.org/how-to-create-tables-in-r/}{Link}.

user2588829, et al.
“Break/Exit Script.” Stack Overflow, 24 July 2013,
\href{https://stackoverflow.com/questions/17837289/break-exit-script}{Link}.
}
