% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoAnalysisPlot.R
\name{mutPlot}
\alias{mutPlot}
\title{Plot The Mutation Results}
\usage{
mutPlot(mutvals, barcolor, title_name, x_name, y_name)
}
\arguments{
\item{mutvals}{A list of mutation detail values of class "mutChecker"
addressing the total number of mutations, the number of Base A mutated,
the number of Base T mutated, the number of Base C mutated,
the number of Base G mutated, the position of each mutation, and
the matrix containing mutation details, respective.}

\item{barcolor}{An optional parameter which determines the bar color.}

\item{title_name}{An optional parameter which determines the title name of
the plot.}

\item{x_name}{An optional parameter which determines x axis label.}

\item{y_name}{An optional parameter which determines y axis label.}
}
\value{
Returns a plot of mutation details.
}
\description{
A function that plots the mutation result details.
}
\examples{
# Example 1:
# Using sampleseq and samplemutseq dataset available with package
\dontrun{
# Read in data
load("./data/sampleseq.rda")
load("./data/samplemutseq.rda")
sampleseq1 <- sampleseq[1:(length(sampleseq))]
samplemutseq1 <- samplemutseq[1:(length(samplemutseq))]

# Check the number of mutations.
mutCheckerResults <- mutChecker(sampleseq1, samplemutseq1)

mutPlot(mutvals = mutCheckerResults)
}

# Example 2:
# Import data in sample raw FASTA format
# Note: to read FASTA files, it requires installation of seqinr R package
library("seqinr")
inputhea <- system.file("extdata", "sample.fasta", package = "oncoAnalysis")
inputmut <- system.file("extdata", "samplemut.fasta", package = "oncoAnalysis")

# Read using read.fasta()
sampleseq <- seqinr::read.fasta(file = inputhea)
samplemutseq <- seqinr::read.fasta(file = inputmut)

# Process and store data
sampleseq1 <- sampleseq$Sample[1:(length(sampleseq$Sample))]
samplemutseq1 <- samplemutseq$Samplemut[1:(length(samplemutseq$Samplemut))]

# Check the number of mutations.
mutCheckerResults <- mutChecker(
                          sampleseq1,
                          samplemutseq1)

mutPlot(mutvals = mutCheckerResults)

}
\references{
Holtz, Yan. “Basic Histogram with GGPLOT2.” – The R Graph Gallery.
\href{https://r-graph-gallery.com/220-basic-ggplot2-histogram.html}{Link}

“Modify Components of a Theme - Theme.” - Theme • ggplot2
\href{https://ggplot2.tidyverse.org/reference/theme.html}{Link}

“Bar Charts - geom_bar.” - geom_bar • ggplot2
\href{https://ggplot2.tidyverse.org/reference/geom_bar.html}{Link}

Wickham, Hadley, Winston Chang, and Maintainer Hadley Wickham.
"Package ‘ggplot2’." Create elegant data visualisations using the grammar
of graphics. Version 2.1 (2016): 1-189.

SimonGSimonG4, et al.
“‘Correct’ Way to Specifiy Optional Arguments in R Functions.”
Stack Overflow, 1 Apr. 1962
\href{https://stackoverflow.com/questions/28370249/correct-way-to-specifiy-optional-arguments-in-r-functions}{Link}

Charif, D. and Lobry, J.R. (2007). SeqinR 1.0-2: a contributed package to the
R project for statistical computing devoted to biological sequences retrieval
and analysis
\href{https://cran.r-project.org/web/packages/seqinr/index.html}{Link}

user2588829, et al.
“Break/Exit Script.” Stack Overflow, 24 July 2013,
\href{https://stackoverflow.com/questions/17837289/break-exit-script}{Link}.
}
