% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutChecker.R
\name{mutTable}
\alias{mutTable}
\title{Display the Mutation Matrix generated by mutChecker in table format}
\usage{
mutTable(mutatmatrix)
}
\arguments{
\item{mutatmatrix}{A mutation matrix generated by mutChecker}
}
\value{
Displays the mutation table.
\itemize{
\item A table containing mutation details. (Column names represent the
base in original sequence, Role names represent the base in
mutated sequence. Used to check for example how many base A has mutated
into base T.)
}
}
\description{
A function that displays Mutation Matrix in table format
}
\examples{
# Example 1:
# Using sampleseq and samplemutseq dataset available with package
\dontrun{
# Read in data
sampleseq1 <- sampleseq[1:(length(sampleseq))]
samplemutseq1 <- samplemutseq[1:(length(samplemutseq))]

# Check the number of mutations.
mutCheckerResults <- mutChecker(
                          sampleseq1,
                          samplemutseq1)
mutTable(mutCheckerResults$MutMatrix)
}

# Example 2:
# Import data in sample raw FASTA format
# Note: to read FASTA files, it requires installation of seqinr R package
library("seqinr")
inputhea <- system.file("extdata", "sample.fasta", package = "oncoAnalysis")
inputmut <- system.file("extdata", "samplemut.fasta", package = "oncoAnalysis")

# Read using read.fasta()
sampleseq <- seqinr::read.fasta(file = inputhea)
samplemutseq <- seqinr::read.fasta(file = inputmut)

# Process and store data
sampleseq1 <- sampleseq$Sample[1:(length(sampleseq$Sample))]
samplemutseq1 <- samplemutseq$Samplemut[1:(length(samplemutseq$Samplemut))]

# Check the number of mutations.
mutCheckerResults <- mutChecker(
                          sampleseq1,
                          samplemutseq1)
mutTable(mutCheckerResults$MutMatrix)

}
\references{
Coghlan, Avril. "Little book of R for Bioinformatics."
Cambridge, UK: Welcome Trust Sanger Institute (2011).
\href{https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/index.html}{Link}.

“How to Create Tables in R?” GeeksforGeeks, 27 Dec. 2021.
\href{https://www.geeksforgeeks.org/how-to-create-tables-in-r/}{Link}.

Charif, D. and Lobry, J.R. (2007). SeqinR 1.0-2: a contributed package to the
R project for statistical computing devoted to biological sequences retrieval
and analysis
\href{https://cran.r-project.org/web/packages/seqinr/index.html}{Link}
}
